refactor: update route structure
move social/[id]/page to social/page
