+++ packages/mesh/app/social/_components/follow-suggestion-feed.tsx
@@ -1,6 +1,6 @@
 import Avatar from '@/components/story-card/avatar'
 
-import type { SuggestedFollowers } from '../[id]/page'
+import type { SuggestedFollowers } from '../page'
 import FollowButton from './follow-button'
 
 export default function FollowSuggestionFeed({
